refactor: numbergenerator 모듈 분리
