+++ src/main/java/lotto/model/Lottos.java
@@ -8,6 +8,8 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
+import lotto.model.numbergenerator.LottoNumberGenerator;
+
 public class Lottos {
     private final List<Lotto> lottos;
