+++ src/main/java/lotto/model/numbergenerator/LottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.numbergenerator;
 
 import java.util.List;
