+++ src/test/java/lotto/model/ShuffleGeneratorTest.java
@@ -5,6 +5,9 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import lotto.model.numbergenerator.LottoNumberGenerator;
+import lotto.model.numbergenerator.ShuffleGenerator;
+
 class ShuffleGeneratorTest {
 
     @Test
